Moved the multiple linear regression class in maestro. Added a mps simple network test for random circuits with measurements in the middle. Fixed an issue with the mps dummy simulator called with a three qubits gate.

+++ Network/SimpleDisconnectedNetwork.h
@@ -2216,7 +2216,8 @@ class SimpleDisconnectedNetwork : public INetwork<Time> {
               lookaheadDepthLocal = layers.size() < 10 || nrQubits <= 10 ? 0
                                     : layers.size() < 20
                                         ? static_cast<int>(lookaheadVal)
-                                    : layers.size() < 35 ? 1.5 * lookaheadVal
+                                       : layers.size() < 35
+                                            ? static_cast<int>(1.5 * lookaheadVal)
                                                          : 2 * lookaheadVal;
             }
